Fixed the bugs 2

+++ src/DataTypes.java
@@ -1,8 +1,8 @@
 public class DataTypes {
     public static long sum(int[] numbers) {
-        int s = 0; // variable to accumulate the sum in!
+        long s = 0; // variable to accumulate the sum in!
         // below is a "foreach" loop which iterates through numbers
-        for (long x : numbers) {
+        for (int x : numbers) {
             s += x;
         }
         return s;
